% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesiantrials-package.R
\docType{package}
\name{bayesiantrials-package}
\alias{bayesiantrials}
\alias{bayesiantrials-package}
\title{bayesiantrials: Bayesian design and analysis of two-arm cluster randomised trials using assurance}
\description{
This package implements the Bayesian approach for design and analysis of two-arm cluster randomised control trial using assurance detailed in Wilson (2022).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://newcastlerse.github.io/bayesiantrials/}
}

}
\keyword{internal}
