% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFunctions.R
\name{BayesInf}
\alias{BayesInf}
\title{Function for performing Bayesian inference using rjags}
\usage{
BayesInf(tot, y, X, C, cl, K, eps = 0.001, modelFilePath = NA)
}
\arguments{
\item{tot}{Total sample size}

\item{y}{Response}

\item{X}{Treatment labels}

\item{C}{Number of clusters}

\item{cl}{Cluster label for each individual}

\item{K}{Number of MCMC samples}

\item{eps}{Analysis Gamma prior parameters}

\item{modelFilePath}{Relative path to JAGS model file}
}
\value{
MCMC samples
}
\description{
Function for performing Bayesian inference using rjags
}
