% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CalculateMode}
\alias{CalculateMode}
\title{Calculate mode of vector}
\usage{
CalculateMode(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
A number
}
\description{
Calculate mode of vector
}
\examples{
CalculateMode(c(1,2,3,2,3,4,2))
}
