% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getMode}
\alias{getMode}
\title{Calculate mode of vector}
\usage{
getMode(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
A number
}
\description{
Calculate mode of vector
}
\examples{
getMode(c(1,2,3,2,3,4,2))
}
