% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFunctions.R
\name{getSampleSize}
\alias{getSampleSize}
\title{Function for finding the required sample size}
\usage{
getSampleSize(
  ss,
  design,
  L,
  K,
  C,
  target = 0.8,
  a = 100,
  eps = 0.001,
  alpha = 0.05,
  model_file_path = NA
)
}
\arguments{
\item{ss}{Vector of sample sizes to check}

\item{design}{Design priors}

\item{L}{Monte Carlo samples}

\item{K}{MCMC Samples}

\item{C}{Cluster Size}

\item{target}{Target assurance}

\item{a}{Dirichlet parameter}

\item{eps}{Analysis Gamma prior parameters}

\item{alpha}{Posterior probability}

\item{model_file_path}{Relative path to JAGS model file}
}
\value{
Minimum sample size for target assurance. Returns target assurance not met for any sample size in ss.
}
\description{
Function for finding the required sample size
}
