% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icons_icc_data}
\alias{icons_icc_data}
\title{Identifying Continence OptioNs after Stroke (ICONS)}
\format{
## `icons_icc_data`
A data frame with 10,001 rows and 2 columns:
\describe{
  \item{V1}{Index}
  \item{V2}{ICC estimate}
}
}
\usage{
icons_icc_data
}
\description{
Estimates of the intra-cluster correlation coefficient based on a
feasibility study \insertRef{thomas2015}{bayesiantrials}
}
\references{
Thomas, Lois H, Beverley French, Christopher J Sutton,
Denise Forshaw, Michael J Leathley, Christopher R Burton,
Brenda Roe, et al. 2015.
“Identifying Continence OptioNs After Stroke (ICONS):
An Evidence Synthesis, Case Study and Exploratory Cluster
Randomised Controlled Trial of the Introduction of a Systematic
Voiding Programme for Patients with Urinary Incontinence After
Stroke in Secondary Care.” Programme Grants for Applied Research 3 (1).
}
\keyword{datasets}
