% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{simulateGaussianCopula}
\alias{simulateGaussianCopula}
\title{Function to simulate from a bivariate Gaussian copula}
\usage{
simulateGaussianCopula(N, rho, m_sigma, s_sigma, gamma)
}
\arguments{
\item{N}{Number of samples}

\item{rho}{Samples from design prior distribution for rho}

\item{m_sigma}{Mean of sigma}

\item{s_sigma}{Stdev of sigma}

\item{gamma}{Correlation parameter}
}
\value{
Joint prior distribution between rho and sigma
}
\description{
Function to simulate from a bivariate Gaussian copula
}
