% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFunctions.R
\name{getAssurance}
\alias{getAssurance}
\title{Function for calculating the assurance}
\usage{
getAssurance(ss, design, L, K, C, a = 100, eps = 0.001, alpha = 0.05)
}
\arguments{
\item{ss}{Sample size}

\item{design}{Design priors}

\item{L}{Monte Carlo Samples}

\item{K}{MCMC Samples}

\item{C}{Number of clusters}

\item{a}{Dirichlet parameter}

\item{eps}{Analysis Gamma prior parameters}

\item{alpha}{Posterior probability}
}
\value{
Assurance
}
\description{
Function for calculating the assurance
}
